refactor: Phase 3.4 — move 5 routers to domain modules (#511)
Migrate telegram, whatsapp, widget, bot_sales, yoomoney_webhook routers
to domain modules. Replace db.integration imports with domain service
imports. Old app/routers/ files become thin facade re-exports.
